Force all VIP sites to use local domain for Photon
The VIP File Service has Photon capabilities, but is served from the
same domain. If the Photon module is enabled, force Jetpack to use
the files service instead of the default Photon domains (`i*.wp.com`)
for internal files.

Externally hosted files continue to use the remote Photon service.

Previously, this was only enabled if you had the VIP_USE_JETPACK_PHOTON
constant enabled. This enables it for all since we want this behavior
everywhere.